+++ buildSrc/src/main/java/Gecko.kt
@@ -9,7 +9,7 @@ object Gecko {
     /**
      * GeckoView Version.
      */
-    const val version = "93.0.20210817094606"
+    const val version = "93.0.20210831093704"
 
     /**
      * GeckoView channel