+++ docs/changelog.md
@@ -11,6 +11,13 @@ permalink: /changelog/
 * [Gecko](https://github.com/mozilla-mobile/android-components/blob/main/buildSrc/src/main/java/Gecko.kt)
 * [Configuration](https://github.com/mozilla-mobile/android-components/blob/main/.config.yml)
 
+* **concept-tabstray**
+  * ⚠️ **This is a breaking change**: This component will be removed in future release.
+     * Instead use the `TabsTray` interface from `browser-tabstray`.
+
+* **browser-state**:
+  * 🌟️ Adds a new `previewImageUrl` in `ContentState` which provides a preview image of the page (e.g. the hero image), if available.
+
 # 94.0.0 (In Development)
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v93.0.0...v94.0.0)
 * [Milestone](https://github.com/mozilla-mobile/android-components/milestone/141?closed=1)